Achievement Awards Demonstration program

Data from a randomized trial of the Achievement Awards Demonstration program, reported in Angrist & Lavy (2009).

Usage

AchievementAwardsRCT

Format

A data frame with 16526 rows and 21 variables:

school_id

Fictitious school identification number

school_type

Factor identifying the school type (Arab religious, Jewish religious, Jewish secular)

pair

Number of treatment pair. Note that 7 is a triple.

treated

Indicator for whether school was in treatment group

year

Cohort year

student_id

Fictitious student identification number

sex

Factor identifying student sex

siblings

Number of siblings

immigrant

Indicator for immigrant status

father_ed

Father's level of education

mother_ed

Mother's level of education

Bagrut_status

Indicator for Bagrut attainment

attempted

Number of Bagrut units attempted

awarded

Number of Bagrut units awarded

achv_math

Indicator for satisfaction of math requirement

achv_english

Indicator for satisfaction of English requirement

achv_hebrew

Indicator for satisfaction of Hebrew requirement

lagscore

Lagged Bagrut score

qrtl

Quartile within distribution of lagscore, calculated by cohort and sex

half

Lower or upper half within distribution of lagscore, calculated by cohort and sex

Source

Angrist Data Archive

References

Angrist, J. D., & Lavy, V. (2009). The effects of high stakes high school achievement awards : Evidence from a randomized trial. American Economic Review, 99(4), 1384-1414. doi:10.1257/aer.99.4.1384