State-level annual mortality rates by cause among 18-20 year-olds

A dataset containing state-level annual mortality rates for select causes of death, as well as data related to the minimum legal drinking age and alcohol consumption.

Usage

MortalityRates

Format

A data frame with 5508 rows and 12 variables:

year

Year of observation

state

identifier for state

count

Number of deaths

pop

Population size

legal

Proportion of 18-20 year-old population that is legally allowed to drink

beertaxa

Beer taxation rate

beerpercap

Beer consumption per capita

winepercap

Wine consumption per capita

spiritpercap

Spirits consumption per capita

totpercap

Total alcohol consumption per capita

mrate

Mortality rate per 10,000

cause

Cause of death

Source

Mastering 'Metrics data archive

References

Angrist, J. D., and Pischke, J. S. (2014). _Mastering'metrics: the path from cause to effect_. Princeton University Press, 2014.

Carpenter, C., & Dobkin, C. (2011). The minimum legal drinking age and public health. _Journal of Economic Perspectives, 25_(2), 133-156. doi:10.1257/jep.25.2.133