Cluster-robust variance-covariance matrix for an lme object.

vcovCR returns a sandwich estimate of the variance-covariance matrix of a set of regression coefficient estimates from a lme object.

# S3 method for lme
vcovCR(obj, cluster, type, target, inverse_var, form = "sandwich", ...)

Arguments

obj
Fitted model for which to calculate the variance-covariance matrix

cluster
Optional expression or vector indicating which observations belong to the same cluster. If not specified, will be set to getGroups(obj).

type
Character string specifying which small-sample adjustment should be used, with available options "CR0", "CR1", "CR1p", "CR1S", "CR2", or "CR3". See "Details" section of vcovCR for further information.

target
Optional matrix or vector describing the working variance-covariance model used to calculate the CR2 and CR4 adjustment matrices. If not specified, the target is taken to be the estimated variance-covariance structure of the lme object.

inverse_var
Optional logical indicating whether the weights used in fitting the model are inverse-variance. If not specified, vcovCR will attempt to infer a value.

form
Controls the form of the returned matrix. The default "sandwich" will return the sandwich variance-covariance matrix. Alternately, setting form = "meat" will return only the meat of the sandwich and setting form = B, where B is a matrix of appropriate dimension, will return the sandwich variance-covariance matrix calculated using B as the bread.

...
Additional arguments available for some classes of objects.

Value

An object of class c("vcovCR","clubSandwich"), which consists of a matrix of the estimated variance of and covariances between the regression coefficient estimates.

See also

Examples

library(nlme)
rat_weight <- lme(weight ~ Time * Diet, data=BodyWeight, ~ Time | Rat) 
vcovCR(rat_weight, type = "CR2")

#>              (Intercept)          Time         Diet2        Diet3    Time:Diet2
#> (Intercept)  20.50523572 -0.0586895628  -20.50523572 -20.50523572  0.0586895628
#> Time         -0.05868956  0.0008918201    0.05868956   0.05868956 -0.0008918201
#> Diet2       -20.50523572  0.0586895628 1192.99481298  20.50523572  0.3728534881
#> Diet3       -20.50523572  0.0586895628   20.50523572 238.25477649 -0.0586895628
#> Time:Diet2    0.05868956 -0.0008918201    0.37285349  -0.05868956  0.0458806483
#> Time:Diet3    0.05868956 -0.0008918201   -0.05868956  -1.87063320  0.0008918201
#>                Time:Diet3
#> (Intercept)  0.0586895628
#> Time        -0.0008918201
#> Diet2       -0.0586895628
#> Diet3       -1.8706331980
#> Time:Diet2   0.0008918201
#> Time:Diet3   0.0237304192

data(egsingle, package = "mlmRev")
math_model <- lme(math ~ year * size + female + black + hispanic, 
                  random = list(~ year | schoolid, ~ 1 | childid), 
                  data = egsingle)
vcovCR(math_model, type = "CR2")

#>               (Intercept)          year          size    femaleMale
#> (Intercept)  1.736519e-02  1.401158e-03 -2.038097e-05 -2.612966e-04
#> year         1.401158e-03  1.731819e-03 -1.427005e-06 -3.077206e-04
#> size        -2.038097e-05 -1.427005e-06  3.502637e-08 -1.233401e-06
#> femaleMale  -2.612966e-04 -3.077206e-04 -1.233401e-06  1.805649e-03
#> black1      -3.864396e-03 -3.597708e-04  5.428152e-07 -4.759591e-04
#> hispanic1   -1.912759e-03 -4.327555e-04 -2.801781e-06 -1.510980e-04
#> year:size   -1.982878e-06 -2.017339e-06  2.488423e-09  2.488713e-07
#>                    black1     hispanic1     year:size
#> (Intercept) -3.864396e-03 -1.912759e-03 -1.982878e-06
#> year        -3.597708e-04 -4.327555e-04 -2.017339e-06
#> size         5.428152e-07 -2.801781e-06  2.488423e-09
#> femaleMale  -4.759591e-04 -1.510980e-04  2.488713e-07
#> black1       5.853760e-03  3.897285e-03  4.954402e-07
#> hispanic1    3.897285e-03  7.563077e-03  5.786971e-07
#> year:size    4.954402e-07  5.786971e-07  2.678030e-09