Cluster-robust variance-covariance matrix for an lmerMod object.

vcovCR returns a sandwich estimate of the variance-covariance matrix of a set of regression coefficient estimates from merMod object.

# S3 method for lmerMod
vcovCR(obj, cluster, type, target, inverse_var, form = "sandwich", ...)

Arguments

obj
Fitted model for which to calculate the variance-covariance matrix

cluster
Optional expression or vector indicating which observations belong to the same cluster. If not specified, will be set to getGroups(obj).

type
Character string specifying which small-sample adjustment should be used, with available options "CR0", "CR1", "CR1p", "CR1S", "CR2", or "CR3". See "Details" section of vcovCR for further information.

target
Optional matrix or vector describing the working variance-covariance model used to calculate the CR2 and CR4 adjustment matrices. If not specified, the target is taken to be the estimated variance-covariance structure of the lmerMod object.

inverse_var
Optional logical indicating whether the weights used in fitting the model are inverse-variance. If not specified, vcovCR will attempt to infer a value.

form
Controls the form of the returned matrix. The default "sandwich" will return the sandwich variance-covariance matrix. Alternately, setting form = "meat" will return only the meat of the sandwich and setting form = B, where B is a matrix of appropriate dimension, will return the sandwich variance-covariance matrix calculated using B as the bread.

...
Additional arguments available for some classes of objects.

Value

An object of class c("vcovCR","clubSandwich"), which consists of a matrix of the estimated variance of and covariances between the regression coefficient estimates.

See also

Examples

library(lme4)

#> 
#> Attaching package: ‘lme4’

#> The following object is masked from ‘package:nlme’:
#> 
#>     lmList

sleep_fit <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
vcovCR(sleep_fit, type = "CR2")

#>             (Intercept)      Days
#> (Intercept)   46.574572 -1.451096
#> Days          -1.451096  2.389463

data(egsingle, package = "mlmRev")
math_model <- lmer(math ~ year * size + female + black + hispanic 
                   + (1 | schoolid) + (1 | childid), 
                   data = egsingle)

#> Warning: Some predictor variables are on very different scales: consider rescaling

vcovCR(math_model, type = "CR2")

#>               (Intercept)          year          size    femaleMale
#> (Intercept)  1.863011e-02  1.594001e-03 -2.150755e-05 -5.152966e-04
#> year         1.594001e-03  1.781083e-03 -2.254868e-06 -2.116067e-04
#> size        -2.150755e-05 -2.254868e-06  3.545270e-08 -9.546631e-07
#> femaleMale  -5.152966e-04 -2.116067e-04 -9.546631e-07  1.874630e-03
#> black1      -3.900798e-03  8.564705e-05  7.891295e-07 -4.272395e-04
#> hispanic1   -2.084991e-03  8.134093e-05 -2.478816e-06 -1.071708e-04
#> year:size   -2.209581e-06 -1.869085e-06  3.494674e-09  1.068855e-07
#>                    black1     hispanic1     year:size
#> (Intercept) -3.900798e-03 -2.084991e-03 -2.209581e-06
#> year         8.564705e-05  8.134093e-05 -1.869085e-06
#> size         7.891295e-07 -2.478816e-06  3.494674e-09
#> femaleMale  -4.272395e-04 -1.071708e-04  1.068855e-07
#> black1       5.612436e-03  3.819343e-03  6.989508e-08
#> hispanic1    3.819343e-03  7.923234e-03 -6.929796e-08
#> year:size    6.989508e-08 -6.929796e-08  2.267422e-09