Cluster-robust variance-covariance matrix for a rma.uni object.

vcovCR returns a sandwich estimate of the variance-covariance matrix of a set of regression coefficient estimates from a rma.uni object.

# S3 method for rma.uni
vcovCR(obj, cluster, type, target, inverse_var, form = "sandwich", ...)

Arguments

obj

Fitted model for which to calculate the variance-covariance matrix

cluster

Expression or vector indicating which observations belong to the same cluster. Required for rma.uni objects.

type

Character string specifying which small-sample adjustment should be used, with available options "CR0", "CR1", "CR1p", "CR1S", "CR2", or "CR3". See "Details" section of vcovCR for further information.

target

Optional matrix or vector describing the working variance-covariance model used to calculate the CR2 and CR4 adjustment matrices. If not specified, the target is taken to be diagonal with entries equal to the estimated marginal variance of the effect sizes.

inverse_var

Optional logical indicating whether the weights used in fitting the model are inverse-variance. If not specified, vcovCR will attempt to infer a value.

form

Controls the form of the returned matrix. The default "sandwich" will return the sandwich variance-covariance matrix. Alternately, setting form = "meat" will return only the meat of the sandwich and setting form = B, where B is a matrix of appropriate dimension, will return the sandwich variance-covariance matrix calculated using B as the bread. form = "estfun" will return the (appropriately scaled) estimating function, the transposed crossproduct of which is equal to the sandwich variance-covariance matrix.

...

Additional arguments available for some classes of objects.

Value

An object of class c("vcovCR","clubSandwich"), which consists of a matrix of the estimated variance of and covariances between the regression coefficient estimates.

See also

Examples

library(metafor)
data(corrdat, package = "robumeta")

mfor_fit <- rma.uni(effectsize ~ males + college + binge,
                     vi = var, data = corrdat, method = "FE")
mfor_fit
#> 
#> Fixed-Effects with Moderators Model (k = 172)
#> 
#> I^2 (residual heterogeneity / unaccounted variability): 71.30%
#> H^2 (unaccounted variability / sampling variability):   3.48
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 168) = 585.3059, p-val < .0001
#> 
#> Test of Moderators (coefficients 2:4):
#> QM(df = 3) = 16.7146, p-val = 0.0008
#> 
#> Model Results:
#> 
#>          estimate      se     zval    pval    ci.lb   ci.ub 
#> intrcpt    0.2199  0.1043   2.1085  0.0350   0.0155  0.4244  * 
#> males     -0.0018  0.0014  -1.3366  0.1813  -0.0045  0.0008    
#> college    0.0373  0.0420   0.8879  0.3746  -0.0451  0.1197    
#> binge      0.1093  0.0431   2.5349  0.0112   0.0248  0.1939  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
mfor_CR2 <- vcovCR(mfor_fit, type = "CR2", cluster = corrdat$studyid)
mfor_CR2
#>               intrcpt         males       college         binge
#> intrcpt  0.0448172280 -4.213246e-04 -1.069754e-02 -0.0224381070
#> males   -0.0004213246  5.532594e-06 -7.974607e-06  0.0001324269
#> college -0.0106975420 -7.974607e-06  3.060450e-02  0.0100417571
#> binge   -0.0224381070  1.324269e-04  1.004176e-02  0.0161097642
coef_test(mfor_fit, vcov = mfor_CR2, test = c("Satterthwaite", "saddlepoint"))
#>    Coef. Estimate      SE t-stat d.f. (Satt) p-val (Satt) Sig.    s.p.
#>  intrcpt  0.21993 0.21170  1.039        8.57        0.327       0.0329
#>    males -0.00182 0.00235 -0.773        9.01        0.459      -0.3020
#>  college  0.03733 0.17494  0.213        6.40        0.838      -8.9009
#>    binge  0.10933 0.12692  0.861       14.01        0.404      -0.1622
#>  p-val (Saddle) Sig.
#>           0.324     
#>           0.456     
#>           0.830     
#>           0.399     
Wald_test(mfor_fit, constraints = constrain_zero(2:4), vcov = mfor_CR2)
#>  test Fstat df_num df_denom p_val sig
#>   HTZ 0.731      3     8.13 0.561