Test all or selected regression coefficients in a fitted model

coef_test reports t-tests for each coefficient estimate in a fitted linear regression model, using a sandwich estimator for the standard errors and a small sample correction for the p-value. The small-sample correction is based on a Satterthwaite approximation or a saddlepoint approximation.

coef_test(
  obj,
  vcov,
  test = "Satterthwaite",
  coefs = "All",
  p_values = TRUE,
  ...
)

Arguments

obj
Fitted model for which to calculate t-tests.

vcov
Variance covariance matrix estimated using vcovCR or a character string specifying which small-sample adjustment should be used to calculate the variance-covariance.

test
Character vector specifying which small-sample corrections to calculate. "z" returns a z test (i.e., using a standard normal reference distribution). "naive-t" returns a t test with m - 1 degrees of freedom. "Satterthwaite" returns a Satterthwaite correction. "saddlepoint" returns a saddlepoint correction. Default is "Satterthwaite".

coefs
Character, integer, or logical vector specifying which coefficients should be tested. The default value "All" will test all estimated coefficients.

p_values
Logical indicating whether to report p-values. The default value is TRUE.

...
Further arguments passed to vcovCR, which are only needed if vcov is a character string.

Value

A data frame containing estimated regression coefficients, standard errors, and test results. For the Satterthwaite approximation, degrees of freedom and a p-value are reported. For the saddlepoint approximation, the saddlepoint and a p-value are reported.

See also

Examples

data("Produc", package = "plm")
lm_individual <- lm(log(gsp) ~ 0 + state + log(pcap) + log(pc) + log(emp) + unemp, data = Produc)
individual_index <- !grepl("state", names(coef(lm_individual)))
coef_test(lm_individual, vcov = "CR2", cluster = Produc$state, coefs = individual_index)

#>       Coef. Estimate     SE t-stat d.f. p-val (Satt) Sig.
#> 1 log(pcap)  -0.0261 0.0625 -0.419 22.9       0.6794     
#> 2   log(pc)   0.2920 0.0646  4.518 22.2       <0.001  ***
#> 3  log(emp)   0.7682 0.0855  8.981 20.4       <0.001  ***
#> 4     unemp  -0.0053 0.0026 -2.040 32.0       0.0497    *

V_CR2 <- vcovCR(lm_individual, cluster = Produc$state, type = "CR2")
coef_test(lm_individual, vcov = V_CR2, coefs = individual_index)

#>       Coef. Estimate     SE t-stat d.f. p-val (Satt) Sig.
#> 1 log(pcap)  -0.0261 0.0625 -0.419 22.9       0.6794     
#> 2   log(pc)   0.2920 0.0646  4.518 22.2       <0.001  ***
#> 3  log(emp)   0.7682 0.0855  8.981 20.4       <0.001  ***
#> 4     unemp  -0.0053 0.0026 -2.040 32.0       0.0497    *