Calculate confidence intervals for all or selected regression coefficients in a fitted model

conf_int reports confidence intervals for each coefficient estimate in a fitted linear regression model, using a sandwich estimator for the standard errors and a small sample correction for the critical values. The small-sample correction is based on a Satterthwaite approximation.

conf_int(
  obj,
  vcov,
  level = 0.95,
  test = "Satterthwaite",
  coefs = "All",
  ...,
  p_values = FALSE
)

Arguments

obj
Fitted model for which to calculate confidence intervals.

vcov
Variance covariance matrix estimated using vcovCR or a character string specifying which small-sample adjustment should be used to calculate the variance-covariance.

level
Desired coverage level for confidence intervals.

test
Character vector specifying which small-sample corrections to calculate. "z" returns a z test (i.e., using a standard normal reference distribution). "naive-t" returns a t test with m - 1 degrees of freedom, where m is the number of unique clusters. "naive-tp" returns a t test with m - p degrees of freedom, where p is the number of regression coefficients in obj. "Satterthwaite" returns a Satterthwaite correction. Unlike in coef_test(), "saddlepoint" is not currently supported in conf_int() because saddlepoint confidence intervals do not have a closed-form solution.

coefs
Character, integer, or logical vector specifying which coefficients should be tested. The default value "All" will test all estimated coefficients.

...
Further arguments passed to vcovCR, which are only needed if vcov is a character string.

p_values
Logical indicating whether to report p-values. The default value is FALSE.

Value

A data frame containing estimated regression coefficients, standard errors, confidence intervals, and (optionally) p-values.

See also

Examples

data("Produc", package = "plm")
lm_individual <- lm(log(gsp) ~ 0 + state + log(pcap) + log(pc) + log(emp) + unemp, data = Produc)
individual_index <- !grepl("state", names(coef(lm_individual)))
conf_int(lm_individual, vcov = "CR2", cluster = Produc$state, coefs = individual_index)
#>      Coef. Estimate     SE d.f. Lower 95% CI Upper 95% CI
#>  log(pcap)  -0.0261 0.0625 22.9      -0.1554     1.03e-01
#>    log(pc)   0.2920 0.0646 22.2       0.1580     4.26e-01
#>   log(emp)   0.7682 0.0855 20.4       0.5900     9.46e-01
#>      unemp  -0.0053 0.0026 32.0      -0.0106    -6.94e-06

V_CR2 <- vcovCR(lm_individual, cluster = Produc$state, type = "CR2")
conf_int(lm_individual, vcov = V_CR2, level = .99, coefs = individual_index)
#>      Coef. Estimate     SE d.f. Lower 99% CI Upper 99% CI
#>  log(pcap)  -0.0261 0.0625 22.9      -0.2016      0.14927
#>    log(pc)   0.2920 0.0646 22.2       0.1099      0.47407
#>   log(emp)   0.7682 0.0855 20.4       0.5253      1.01102
#>      unemp  -0.0053 0.0026 32.0      -0.0124      0.00182