Detect cluster structure of an rma.mv object

findCluster.rma.mv returns a vector of ID variables for the highest level of clustering in a fitted rma.mv model.

Usage

findCluster.rma.mv(obj)

Arguments

obj

A fitted rma.mv object.

Value

A a vector of ID variables for the highest level of clustering in obj.

Examples

library(metafor)
#> Loading required package: Matrix
#> Loading required package: metadat
#> 
#> Loading the 'metafor' package (version 3.4-0). For an
#> introduction to the package please type: help(metafor)
data(hierdat, package = "robumeta")

mfor_fit <- rma.mv(effectsize ~ binge + followup + sreport + age, 
                 V = var, random = list(~ 1 | esid, ~ 1 | studyid),
                 data = hierdat)
findCluster.rma.mv(mfor_fit)
#>  [1] 1  1  1  1  2  2  2  2  3  3  4  4  5  5  5  5  5  5  5  6  6  7  7  7  8 
#> [26] 8  8  9  9  9  9  9  10 11 12 12 13 13 14 15 15 15 15 15 15 15 15 15 15 15
#> [51] 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15
#> Levels: 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15