Cluster-robust variance-covariance matrix for a robu object.

vcovCR returns a sandwich estimate of the variance-covariance matrix of a set of regression coefficient estimates from a rma.mv object.

# S3 method for rma.mv
vcovCR(obj, cluster, type, target, inverse_var, form = "sandwich", ...)

Arguments

obj
Fitted model for which to calculate the variance-covariance matrix

cluster
Optional expression or vector indicating which observations belong to the same cluster. If not specified, will be set to the factor in the random-effects structure with the fewest distinct levels. Caveat emptor: the function does not check that the random effects are nested.

type
Character string specifying which small-sample adjustment should be used, with available options "CR0", "CR1", "CR1p", "CR1S", "CR2", or "CR3". See "Details" section of vcovCR for further information.

target
Optional matrix or vector describing the working variance-covariance model used to calculate the CR2 and CR4 adjustment matrices. If not specified, the target is taken to be the estimated variance-covariance structure of the rma.mv object.

inverse_var
Optional logical indicating whether the weights used in fitting the model are inverse-variance. If not specified, vcovCR will attempt to infer a value.

form
Controls the form of the returned matrix. The default "sandwich" will return the sandwich variance-covariance matrix. Alternately, setting form = "meat" will return only the meat of the sandwich and setting form = B, where B is a matrix of appropriate dimension, will return the sandwich variance-covariance matrix calculated using B as the bread. form = "estfun" will return the (appropriately scaled) estimating function, the transposed crossproduct of which is equal to the sandwich variance-covariance matrix.

...
Additional arguments available for some classes of objects.

Value

An object of class c("vcovCR","clubSandwich"), which consists of a matrix of the estimated variance of and covariances between the regression coefficient estimates.

See also

Examples

library(metafor)
data(hierdat, package = "robumeta")

mfor_fit <- rma.mv(effectsize ~ binge + followup + sreport + age, 
                 V = var, random = list(~ 1 | esid, ~ 1 | studyid),
                 data = hierdat)
mfor_fit
#> 
#> Multivariate Meta-Analysis Model (k = 68; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed   factor 
#> sigma^2.1  0.1195  0.3457     68     no     esid 
#> sigma^2.2  0.0000  0.0000     15     no  studyid 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 63) = 229.6741, p-val < .0001
#> 
#> Test of Moderators (coefficients 2:5):
#> QM(df = 4) = 47.5517, p-val < .0001
#> 
#> Model Results:
#> 
#>           estimate      se     zval    pval    ci.lb   ci.ub 
#> intrcpt     0.3898  0.6658   0.5855  0.5582  -0.9151  1.6946      
#> binge       0.4470  0.1251   3.5717  0.0004   0.2017  0.6922  *** 
#> followup    0.0013  0.0009   1.5534  0.1203  -0.0003  0.0030      
#> sreport     0.5385  0.1319   4.0839  <.0001   0.2801  0.7970  *** 
#> age        -0.0433  0.0402  -1.0772  0.2814  -0.1222  0.0355      
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

mfor_CR2 <- vcovCR(mfor_fit, type = "CR2")
mfor_CR2
#>                intrcpt         binge      followup       sreport           age
#> intrcpt   0.4242620752 -7.862569e-03 -2.876222e-04 -1.720867e-03 -2.400533e-02
#> binge    -0.0078625691  1.033451e-02 -1.569351e-05  3.872261e-03  3.644227e-04
#> followup -0.0002876222 -1.569351e-05  5.179354e-07 -1.949264e-05  1.639318e-05
#> sreport  -0.0017208671  3.872261e-03 -1.949264e-05  2.039661e-02 -8.265205e-04
#> age      -0.0240053252  3.644227e-04  1.639318e-05 -8.265205e-04  1.403472e-03
coef_test(mfor_fit, vcov = mfor_CR2, test = c("Satterthwaite", "saddlepoint"))
#>     Coef. Estimate      SE t-stat d.f. (Satt) p-val (Satt) Sig.    s.p.
#>   intrcpt  0.38977 0.65135  0.598        3.07       0.5908      -0.6612
#>     binge  0.44697 0.10166  4.397        3.59       0.0148    *  0.4162
#>  followup  0.00133 0.00072  1.852        2.03       0.2030       0.2552
#>   sreport  0.53855 0.14282  3.771        4.35       0.0168    *  0.4114
#>       age -0.04332 0.03746 -1.156        2.71       0.3392       0.0932
#>  p-val (Saddle) Sig.
#>         0.59591     
#>         0.00451   **
#>         0.18683     
#>         0.00755   **
#>         0.34384     

Wald_test(mfor_fit, constraints = constrain_zero(c(2,4)), vcov = mfor_CR2)
#>  test Fstat df_num df_denom  p_val sig
#>   HTZ  10.3      2     3.47 0.0345   *
Wald_test(mfor_fit, constraints = constrain_zero(2:5), vcov = mfor_CR2)
#>  test Fstat df_num df_denom p_val sig
#>   HTZ 0.177      4   0.0396 0.961